Refactor Expo project with modular layout
