+++ expo-app/app/index.tsx
@@ -1,21 +1,5 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import { StatusBar } from 'expo-status-bar';
+import HomeScreen from '@screens/HomeScreen';
 
-export default function HomeScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>Open up app/index.tsx to start working on your app!</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+export default function Index() {
+  return <HomeScreen />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
