+++ expo-app/package.json
@@ -19,7 +19,8 @@
     "@types/react": "~18.2.14",
     "expo-router": "^2.0.0",
     "react-native-safe-area-context": "^4.7.0",
-    "react-native-screens": "^3.22.0"
+    "react-native-screens": "^3.22.0",
+    "react-native-gesture-handler": "^2.12.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
