+++ expo-app/src/screens/HomeScreen.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+import { StatusBar } from 'expo-status-bar';
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Open up app/index.tsx to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
